service: Split GrpcChannelPool into a separate submodule
